restructure Progress page
